+++ packages/shakedown_core/lib/ui/widgets/settings/tv_screensaver_section.dart
@@ -13,6 +13,7 @@ import 'package:shakedown_core/ui/widgets/tv/tv_list_tile.dart';
 import 'package:shakedown_core/ui/widgets/tv/tv_stepper_row.dart';
 
 part 'tv_screensaver_section_build.dart';
+part 'tv_screensaver_section_audio_build.dart';
 part 'tv_screensaver_section_controls.dart';
 
 class TvScreensaverSection extends StatefulWidget {